% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom.cluster.medians.R
\name{fsom.cluster.medians}
\alias{fsom.cluster.medians}
\title{fsom.cluster.medians}
\usage{
fsom.cluster.medians(fsom, dat.type = c("map", "data", "codes"), ranged = TRUE)
}
\value{

}
\description{
fsom.cluster.medians
}
