% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom.explorer.R
\name{fsom.explorer}
\alias{fsom.explorer}
\title{fsom.explorer}
\usage{
fsom.explorer(
  input.fsom,
  name.X1,
  name.Y1,
  name.X2,
  name.Y2,
  name.X3,
  name.Y3,
  points,
  ...
)
}
\value{

}
\description{
fsom.explorer
}
