% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsom.heatmap.R
\name{fsom.heatmap}
\alias{fsom.heatmap}
\title{fsom.heatmap}
\usage{
fsom.heatmap(
  fsom,
  dat.type,
  dims,
  row_order = NULL,
  clust.freq = T,
  clust.count = F,
  display_numbers = FALSE,
  color.cut = 100,
  color.option = "brewer",
  ...
)
}
\value{

}
\description{
fsom.heatmap
}
