% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStars.edit.R
\name{PlotStars.edit}
\alias{PlotStars.edit}
\title{PlotStars.edit}
\usage{
PlotStars.edit(
  median.vals,
  fsom,
  markers = fsom$map$colsUsed,
  view = "MST",
  colorPalette = grDevices::colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
    "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")),
  starBg = "white",
  backgroundValues = NULL,
  backgroundColor = function(n) {     grDevices::rainbow(n, alpha = 0.3) },
  backgroundLim = NULL,
  backgroundBreaks = NULL,
  backgroundSize = NULL,
  thresholds = NULL,
  legend = TRUE,
  query = NULL,
  main = ""
)
}
\value{

}
\description{
PlotStars.edit
}
